master can now also handle databases; refactor.
