+++ src/atc/__init__.py
@@ -5,4 +5,4 @@ See https://github.com/atc-net/atc-dataplatform for details
 __version__ = "0.1.5"
 
 from .spark import Spark
-from .config_master import ConfigMaster
+from .config_master.config_master import ConfigMaster
