+++ src/atc/config_master/exceptions.py
@@ -1,0 +1,10 @@
+class SqlParseException(Exception):
+    pass
+
+
+class NoCreationException(SqlParseException):
+    pass
+
+
+class UnknownCreationException(SqlParseException):
+    pass
