+++ tests/config_master/sql/db1/db1.sql
@@ -1,3 +1,3 @@
-CREATE DATABASE IF NOT EXISTS my_db1
+CREATE DATABASE IF NOT EXISTS "my_db1{ID}"
 COMMENT "Dummy Database 1"
 LOCATION "/tmp/foo/bar/my_db1/"
